+++ BuildMetrics.Lib/Interfaces/IConcrete.cs
@@ -6,5 +6,5 @@ public interface IConcrete
     /// Get volume of concrete required for construction
     /// </summary>
     /// <returns>Volume of concrete</returns>
-    float GetVolumeOfConcrete();
+    double GetVolumeOfConcrete();
 }
