+++ BuildMetrics.Lib/Interfaces/IExcavation.cs
@@ -6,11 +6,11 @@ public interface IExcavation
     /// Get volume of earth to be excavated
     /// </summary>
     /// <returns>Volume of excavation</returns>
-    float GetVolumeOfExcavation();
+    double GetVolumeOfExcavation();
 
     /// <summary>
     /// Get volume of earth to be carted away
     /// </summary>
     /// <returns>Volume of cart away</returns>
-    float GetVolumeOfCartAway();
+    double GetVolumeOfCartAway();
 }
